Require token_filter and use site:url in the href attribute
The installer used sed to make changes to the href
attribute to prepend a site prefix. Rather than doing that,
require the token_filter, and use the site:url token
as the beginning of the href attribute.

NB: now also require the x1 module, which defines the
basic_html_token format.